Adds new option for simulating a headless session
Some runtimes which don't support `XR_MND_headless` are able to run as if they do implement this extension (e.g. Oculus PC runtime). This commit adds an option to first try enabling a real headless extension and if not available will attempt to run as if one is supported, a graphics context is still created but no swapchains are created and frames are not submitted.